dom/html/test/imports/test_defaultView.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1061469
-->
<head>
  <title>Test for Bug 1061469</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta http-equiv="Content-Type" content="text/html;charset=utf-8">
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1061469">Mozilla Bug 1061469</a>
  <script type="text/javascript">
    SimpleTest.waitForExplicitFinish();
    var success = false;
    function loaded() {
      success = true;
    }
    function failed() {
      ok(false, "Import loading failed");
    }
  </script>
  <link rel="import" href="file_simple_import.html" id="import" onload="loaded()" onerror="failed()"></link>
  <script type="text/javascript">
    document.defaultView;
    is(document.getElementById("import").import.defaultView, null, "defaultView is always null for imports");
    SimpleTest.finish();
  </script>
</body>
</html>
